Push to GitHub Initially
It is not yet done. I am completing the project. Currently, it shows the total grade histogram and other statistics like mode, mean, etc. I am working on an exam that has n questions. There are the first column for the IDs and the n columns for each question which are 0 or 1 and a total column.
The current accepted file is Excel file.